+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Forge.java
@@ -17,55 +17,56 @@ public class Forge implements ICommand<MessageEvent<Bot>>
 	@Override
 	public void exe(MessageEvent<Bot> event) throws IOException, IncorrectCommandExecutionException
 	{
-		String[] args = Utilities.toArgs(event.getMessage());
-
-		if(args.length == 3)
-		{
-			if(args[1].equalsIgnoreCase("latest") || args[1].equalsIgnoreCase("latets"))
-			{
-				if(args[2].equalsIgnoreCase("version"))
-				{
-					String result = createResult("<td>1.8-Latest</td>", "Version");
-
-					if(result.equalsIgnoreCase(createResult("<td>1.8-Recommended</td>", "Version")))
-						Utilities.chanMsg(event, result + " " + L10N.getString("forge.latest.recommended"));
-					else
-						Utilities.chanMsg(event, result);
-				}
-				else if(args[2].equalsIgnoreCase("changelog"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Changelog"));
-				else if(args[2].equalsIgnoreCase("dlmain"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Installer"));
-				else if(args[2].equalsIgnoreCase("dlsrc"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Src"));
-				else
-					throw new IncorrectCommandExecutionException(getAlias());
-			}
-			else if(args[1].equalsIgnoreCase("rec") || args[1].equalsIgnoreCase("recommended"))
-			{
-				if(args[2].equalsIgnoreCase("version"))
-				{
-					String result = createResult("<td>1.8-Recommended</td>", "Version");
-
-					if(result.equalsIgnoreCase(createResult("<td>Latest-1.8</td>", "Version")))
-						Utilities.chanMsg(event, result + " " + L10N.getString("forge.recommended.latest"));
-					else
-						Utilities.chanMsg(event, result);
-				}
-				else if(args[2].equalsIgnoreCase("changelog"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Changelog"));
-				else if(args[2].equalsIgnoreCase("dlmain"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Installer"));
-				else if(args[2].equalsIgnoreCase("dlsrc"))
-					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Src"));
-				else
-					throw new IncorrectCommandExecutionException(getAlias());
-			}
-			else
-				throw new IncorrectCommandExecutionException(getAlias());
-		}
-		else
-			throw new IncorrectCommandExecutionException(getAlias());
+		Utilities.chanMsg(event, L10N.getString("forge.disabled"));
+//		String[] args = Utilities.toArgs(event.getMessage());
+//
+//		if(args.length == 3)
+//		{
+//			if(args[1].equalsIgnoreCase("latest") || args[1].equalsIgnoreCase("latets"))
+//			{
+//				if(args[2].equalsIgnoreCase("version"))
+//				{
+//					String result = createResult("<td>1.8-Latest</td>", "Version");
+//
+//					if(result.equalsIgnoreCase(createResult("<td>1.8-Recommended</td>", "Version")))
+//						Utilities.chanMsg(event, result + " " + L10N.getString("forge.latest.recommended"));
+//					else
+//						Utilities.chanMsg(event, result);
+//				}
+//				else if(args[2].equalsIgnoreCase("changelog"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Changelog"));
+//				else if(args[2].equalsIgnoreCase("dlmain"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Installer"));
+//				else if(args[2].equalsIgnoreCase("dlsrc"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Latest</td>", "Src"));
+//				else
+//					throw new IncorrectCommandExecutionException(getAlias());
+//			}
+//			else if(args[1].equalsIgnoreCase("rec") || args[1].equalsIgnoreCase("recommended"))
+//			{
+//				if(args[2].equalsIgnoreCase("version"))
+//				{
+//					String result = createResult("<td>1.8-Recommended</td>", "Version");
+//
+//					if(result.equalsIgnoreCase(createResult("<td>Latest-1.8</td>", "Version")))
+//						Utilities.chanMsg(event, result + " " + L10N.getString("forge.recommended.latest"));
+//					else
+//						Utilities.chanMsg(event, result);
+//				}
+//				else if(args[2].equalsIgnoreCase("changelog"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Changelog"));
+//				else if(args[2].equalsIgnoreCase("dlmain"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Installer"));
+//				else if(args[2].equalsIgnoreCase("dlsrc"))
+//					Utilities.chanMsg(event, createResult("<td>1.8-Recommended</td>", "Src"));
+//				else
+//					throw new IncorrectCommandExecutionException(getAlias());
+//			}
+//			else
+//				throw new IncorrectCommandExecutionException(getAlias());
+//		}
+//		else
+//			throw new IncorrectCommandExecutionException(getAlias());
 	}
 
 	private String createResult(String type, String action) throws IOException